% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indicator.R
\name{convert_to_long_format}
\alias{convert_to_long_format}
\title{Convert wide format data to long format}
\usage{
convert_to_long_format(df, config)
}
\arguments{
\item{df}{Data frame in wide format}

\item{config}{Indicator configuration}
}
\value{
Data frame in long format
}
\description{
Convert wide format data to long format
}
