% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_web_scraping.R
\name{find_excel_in_page}
\alias{find_excel_in_page}
\title{Search for Excel files within a specific page}
\usage{
find_excel_in_page(page_url)
}
\arguments{
\item{page_url}{URL to search for Excel files}
}
\value{
Excel file URL or NA
}
\description{
Search for Excel files within a specific page
}
