% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indicator.R
\name{process_excel_file}
\alias{process_excel_file}
\title{Process Excel file according to indicator configuration}
\usage{
process_excel_file(excel_file, config)
}
\arguments{
\item{excel_file}{Path to Excel file}

\item{config}{Indicator configuration list}
}
\value{
Processed data frame
}
\description{
Process Excel file according to indicator configuration
}
